+++ src/use-async-combine-all.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/named
-import { UseAsyncCombine } from './index';
-
-export const useAsyncCombineAll: UseAsyncCombine;
-export default useAsyncCombineAll;