+++ src/use-async-combine-race.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/named
-import { UseAsyncCombine } from './index';
-
-export const useAsyncCombineRace: UseAsyncCombine;
-export default useAsyncCombineRace;