+++ src/use-async-combine-seq.d.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/named
-import { UseAsyncCombine } from './index';
-
-export const useAsyncCombineSeq: UseAsyncCombine;
-export default useAsyncCombineSeq;